move macOS code into Mio folder
