+++ Mio/Shared Views/ImagePicker.swift
@@ -1,0 +1,22 @@
+//
+//  ImagePicker.swift
+//  Mio
+//
+//  Created by Finn Behrens on 13.06.21.
+//  Copyright © 2021 Kilian Koeltzsch. All rights reserved.
+//
+
+import SwiftUI
+
+struct ImagePicker: View {
+
+    var body: some View {
+        Text("Sorry, no image picker on macOS yet :-/")
+    }
+}
+
+struct ImagePicker_Previews: PreviewProvider {
+    static var previews: some View {
+        ImagePicker()
+    }
+}
