+++ Nio/Shared Views/ImagePicker.swift
@@ -1,15 +1,5 @@
 import SwiftUI
 
-#if os(macOS)
-
-struct ImagePicker: View {
-  
-    var body: some View {
-        Text("Sorrz, no image picker on macOS yet :-/")
-    }
-}
-
-#else // iOS
 struct ImagePicker: UIViewControllerRepresentable {
 
     @Environment(\.presentationMode)
@@ -70,4 +60,3 @@ struct ImagePicker: UIViewControllerRepresentable {
     }
 
 }
-#endif // iOS
