+++ Nio/Shared Views/MessageTextViewWrapper.swift
@@ -1,47 +1,7 @@
 import SwiftUI
 import NioKit
 
-#if os(macOS)
-class MessageTextView: NSTextView {
-    convenience init(attributedString: NSAttributedString, linkColor: UXColor,
-                     maxSize: CGSize)
-    {
-        self.init()
-        backgroundColor = .clear
-        textContainerInset = .zero
-        isEditable = false
-        linkTextAttributes = [
-            .foregroundColor: linkColor,
-            .underlineStyle: NSUnderlineStyle.single.rawValue,
-        ]
 
-        self.insertText(attributedString,
-                        replacementRange: NSRange(location: 0, length: 0))
-        self.maxSize = maxSize
-
-        // don't resist text wrapping across multiple lines
-        setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
-    }
-}
-
-struct MessageTextViewWrapper: NSViewRepresentable {
-    let attributedString: NSAttributedString
-    let linkColor: NSColor
-    let maxSize: CGSize
-
-    func makeNSView(context: Context) -> MessageTextView {
-        MessageTextView(attributedString: attributedString, linkColor: linkColor, maxSize: maxSize)
-    }
-
-    func updateNSView(_ uiView: MessageTextView, context: Context) {
-        // nothing to update
-    }
-
-    func makeCoordinator() {
-        // nothing to coordinate
-    }
-}
-#else // iOS
 /// An automatically sized label, which allows links to be tapped.
 class MessageTextView: UITextView {
     var maxSize: CGSize = .zero
@@ -92,4 +52,3 @@ struct MessageTextViewWrapper: UIViewRepresentable {
         // nothing to coordinate
     }
 }
-#endif // iOS
